requirements.txt
spacy
streamlit
PyPDF2
python-docx
fuzzywuzzy
python-Levenshtein
